Have emerge check for already installed packages
This avoids asking for root privileges every time autoproj is used with
"osdeps_mode" configuration including "os".

emerges package names(atoms) are complicated enough that only the simplest
cases can be resolved by going to the package database. Checking for
updates requires parsing all the available repositories, cross-checking
use flags and so on, which is kept strictly with emerge. Luckily, that
is only required for the osdeps command.